Remove tekton
Tekton was used to clone a git repository in an init container. This
is replaced  by an internal tool called gitcloner that uses go-git to clone the
repositories.

gitcloner is a cli tool that clones a specific branch of a git
repository.

GitJob uses gitcloner in an init-container to clone a git repository
into a emptyDir volume. Then the job container uses this emptyDir volume
that contains the repo cloned