Move from wrangler to controller-runtime
Move the gitjob controller from wrangler to controller-runtime.

GitJob controller watch for changes in the GitJob resources and the Jobs owned by them.

Use k8s owner references for the Jobs created in the GitJob reconciliation loop. This allows k8s to garbage collect the Jobs when the GitJob owner is deleted.

Extract the git repository polling outside of the reconciliation loop. Each repository have a watch that fetches the latest commit every syncInterval in a separate goroutine